+++ R/zzz.R
@@ -1,2 +1,3 @@
 # Reexported helpers used in the plot method
-utils::globalVariables(c("time", "median", "lower", "upper", "group"))
+utils::globalVariables(c("time", "median", "lower", "upper", "group", "surv",
+                         "posterior_mean", "posterior_lwr", "posterior_upr"))